requirements.txt
amqp==5.0.2
asgiref==3.3.1
billiard==3.6.3.0
celery==5.0.4
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
Django==3.1.4
djangorestframework==3.12.2
et-xmlfile==1.0.1
httplib2==0.18.1
idna==2.10
importlib-metadata==3.1.1
jdcal==1.4.1
kombu==5.0.2
openpyxl==3.0.5
prompt-toolkit==3.0.8
pytz==2020.4
redis==3.5.3
requests==2.25.0
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.2
vine==5.0.0
wcwidth==0.2.5
zipp==3.4.0
